Migrate most of the launcher code to tasks.
GestorLauncher will still remain as a class.
