+++ src/main/kotlin/com/gestormc/gestor/launcher/GestorLauncher.kt
@@ -1,10 +1,7 @@
 package com.gestormc.gestor.launcher
 
 import com.gestormc.gestor.data.ManifestReleaseType
-import com.gestormc.gestor.launcher.core.ArgumentManager
-import com.gestormc.gestor.launcher.core.AuthManager
-import com.gestormc.gestor.launcher.core.LibraryManager
-import com.gestormc.gestor.launcher.core.SetupManager
+import com.gestormc.gestor.launcher.core.*
 import com.gestormc.gestor.launcher.fabric.FabricLauncherPlugin
 import com.gestormc.gestor.launcher.forge.ForgeLauncherPlugin
 import com.gestormc.gestor.util.InternalAPI
@@ -23,7 +20,8 @@ class GestorLauncher private constructor(
     private val isServer: Boolean,
     private val plugins: MutableSet<LauncherPlugin> = mutableSetOf(),
     private val authentication: YggdrasilUserAuthentication? = null,
-    private val jarTemplate: String = if (isServer) "server" else "client"
+    private val jarTemplate: String = if (isServer) "server" else "client",
+    private val data: LibraryData = LibraryData()
 ) {
 
     /**
@@ -62,11 +60,11 @@ class GestorLauncher private constructor(
 
         plugins.forEach { plugin -> plugin.onSetupStart(root, version, optInLegacyJava) }
 
-        SetupManager.setupVersionInfo(root, version)
-        SetupManager.setupLibraries(root, version)
-        SetupManager.setupJAR(root, version, isServer)
-        SetupManager.setupJava(optInLegacyJava)
-        SetupManager.setupAssets(root, version)
+        launcherVersionInfoSetupTask(root, version)
+        launcherSetupLibrariesTask(root, version, data)
+        launcherSetupJarTask(root, version, isServer)
+        launcherSetupJavaTask(optInLegacyJava)
+        launcherSetupAssetsTask(root, version)
 
         // Make dirs
         File("$root/assets").mkdirs()
@@ -145,7 +143,7 @@ class GestorLauncher private constructor(
 
         // Generate arguments
         var arguments = if (versionInfoObject.contains("minecraftArguments")) {
-            ArgumentManager.generateLegacyArguments(
+            launcherLegacyArgumentsTask(
                 raw = versionInfoObject["minecraftArguments"]!!.jsonPrimitive.content,
                 root = root,
                 version = version,
@@ -156,7 +154,7 @@ class GestorLauncher private constructor(
                     versionInfoObject["assets"]!!.jsonPrimitive.content
             )
         } else {
-            ArgumentManager.generateModernArguments(
+            launcherModernArgumentsTask(
                 version = version,
                 root = root,
                 assetsIndexName =
@@ -182,7 +180,7 @@ class GestorLauncher private constructor(
             )
         }
 
-        var jvmArguments = ArgumentManager.generateJVMArguments(maxMemory.toString(), jvmArgs)
+        var jvmArguments = launcherJVMArgumentsTask(maxMemory.toString(), jvmArgs)
         plugins.forEach { plugin ->
             jvmArguments = plugin.processJvmArguments(
                 jvmArguments,
@@ -228,15 +226,10 @@ class GestorLauncher private constructor(
         }
 
         var classpath = ".;$root/versions/$version/$version-$jarTemplate.jar;${
-            LibraryManager.getLibrariesFormatted(
-                root,
-                versionInfoObject,
-                replacers,
-                exceptions
-            )
+            launcherLibraryFormatTask(root, versionInfoObject, replacers, exceptions, data)
         }"
-        if (versionInfoObject.contains("inheritsFrom")) { // inheritance support
-            classpath += LibraryManager.getLibrariesFormatted(root, getParentObject(versionInfoObject, root))
+        if (versionInfoObject.contains("inheritsFrom")) { // inheritance support. No replacers or exceptions are applied here
+            classpath += launcherLibraryFormatTask(root, getParentObject(versionInfoObject, root), data = data)
         }
         plugins.forEach { plugin ->
             classpath = plugin.processClasspath(
@@ -358,10 +351,9 @@ class GestorLauncher private constructor(
              * Do **not** set this to `true` outside of testing!
              */
             testingLaunch: Boolean = false
-        )
+        ): GestorLauncher {
 
-                : GestorLauncher {
-            val auth = if (testingLaunch) null else AuthManager.start()
+            val auth = if (testingLaunch) null else launcherAuthenticateTask()
             auth?.logIn()
             return GestorLauncher(root, isServer, authentication = auth)
         }
@@ -386,7 +378,7 @@ class GestorLauncher private constructor(
             testingLaunch: Boolean = false
         ): GestorLauncher {
 
-            val auth = if (testingLaunch) null else AuthManager.start()
+            val auth = if (testingLaunch) null else launcherAuthenticateTask()
             auth?.logIn()
             return GestorLauncher(root, isServer, authentication = auth).withPlugin(FabricLauncherPlugin)
         }
@@ -411,7 +403,7 @@ class GestorLauncher private constructor(
             testingLaunch: Boolean = false
         ): GestorLauncher {
 
-            val auth = if (testingLaunch) null else AuthManager.start()
+            val auth = if (testingLaunch) null else launcherAuthenticateTask()
             auth?.logIn()
             return GestorLauncher(root, isServer, authentication = auth).withPlugin(ForgeLauncherPlugin)
         }
